Add BaseServiceTest class
Edit TaskServiceTest to extends BaseSer* class
Inject Dependency for Task modules
